Support local logging via standard logging interface
Change-type: patch
