+++ src/logging-backends/index.ts
@@ -1,0 +1,10 @@
+import { LocalLogBackend } from './local-backend';
+import { LogBackend, LogMessage } from './log-backend';
+import { ResinLogBackend } from './resin-backend';
+
+export {
+	LocalLogBackend,
+	LogBackend,
+	LogMessage,
+	ResinLogBackend,
+};
