+++ src/supervisor.coffee
@@ -22,6 +22,7 @@ startupConfigFields = [
 	'mixpanelToken'
 	'mixpanelHost'
 	'loggingEnabled'
+	'localMode'
 ]
 
 module.exports = class Supervisor extends EventEmitter
@@ -57,7 +58,13 @@ module.exports = class Supervisor extends EventEmitter
 					deviceApiKey: conf.deviceApiKey,
 					offlineMode: checkTruthy(conf.offlineMode),
 					enableLogs: checkTruthy(conf.loggingEnabled),
+					localMode: checkTruthy(conf.localMode)
 				})
+
+				# Setup log backend switching for local mode changes
+				@config.on 'change', (changed) =>
+					if changed.localMode?
+						@logger.switchBackend(checkTruthy(changed.localMode))
 			.then =>
 				@deviceState.init()
 			.then =>
